i am good at programming

+++ math2.h
@@ -1,14 +1,14 @@
 #ifndef __MATH2H__
 #define __MATH2H__
-float degtorad(double degtoconv);
+float degtorad(float degtoconv);
 
-float radtodeg(double radtoconv);
+float radtodeg(float radtoconv);
 
 float point_direction(int x1, int y1, int x2, int y2);
 
-float dsin(double dsinarg);
+float dsin(float dsinarg);
 
-float dcos(double dcosarg);
+float dcos(float dcosarg);
 
 float point_distance(x1,y1,x2,y2);
 #endif
